Code refractored to MVP model from MVC model
