+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/MyMaterialTheme">
-        <activity android:name=".IncomingCallActivity"></activity>
+        <activity android:name=".presenter.IncomingCallActivity"></activity>
 
         <receiver android:name=".CallReceiver" >
             <intent-filter>
@@ -25,8 +25,8 @@
                 <action android:name="android.intent.action.NEW_OUTGOING_CALL" />
             </intent-filter>
         </receiver>
-        <activity android:name=".ContactActivity"></activity>
-        <activity android:name=".MainActivity">
+        <activity android:name=".view.ContactActivity"></activity>
+        <activity android:name=".view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
