+++ app/src/main/java/com/themaskedbit/tempcontact/presenter/BasePresenter.java
@@ -1,0 +1,4 @@
+package com.themaskedbit.tempcontact.presenter;
+
+public interface BasePresenter {
+}
