+++ app/src/main/java/com/themaskedbit/tempcontact/view/BaseView.java
@@ -1,0 +1,5 @@
+package com.themaskedbit.tempcontact.view;
+
+public interface BaseView<T> {
+    void setPresenter(T presenter);
+}
