+++ app/src/main/java/com/themaskedbit/tempcontact/view/TempFragment.java
@@ -1,58 +1,42 @@
-package com.themaskedbit.tempcontact;
+package com.themaskedbit.tempcontact.view;
 
-import android.app.SearchManager;
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v7.widget.SearchView;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.LayoutInflater;
-import android.view.Menu;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import com.themaskedbit.tempcontact.R;
+import com.themaskedbit.tempcontact.TempContract;
+import com.themaskedbit.tempcontact.model.Contact;
 
 import java.util.ArrayList;
 import java.util.List;
 
 
-/**
- * A simple {@link Fragment} subclass.
- * Activities that contain this fragment must implement the
- * {@link TempFragment.OnFragmentInteractionListener} interface
- * to handle interaction events.
- * Use the {@link TempFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
-public class TempFragment extends Fragment {
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+
+public class TempFragment extends Fragment implements TempContract.View {
     private static final String ARG_PARAM1 = "param1";
     private static final String ARG_PARAM2 = "param2";
-
     // TODO: Rename and change types of parameters
     private String mParam1;
     private String mParam2;
     private TextView textView;
-
     private OnFragmentInteractionListener mListener;
+    private TempContract.Presenter presenter;
 
     public TempFragment() {
         // Required empty public constructor
     }
 
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment TempFragment.
-     */
+
     // TODO: Rename and change types and number of parameters
     public static TempFragment newInstance(String param1, String param2) {
         TempFragment fragment = new TempFragment();
@@ -87,7 +71,7 @@ public class TempFragment extends Fragment {
         textView = (TextView) getActivity().findViewById(R.id.action_search);
         final List<Contact> ContactList = new ArrayList<>();
         final CallLogAdapter mAdapter = new CallLogAdapter(ContactList);
-       textView.addTextChangedListener(new TextWatcher() {
+        textView.addTextChangedListener(new TextWatcher() {
            @Override
            public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
 
@@ -100,25 +84,15 @@ public class TempFragment extends Fragment {
 
            @Override
            public void afterTextChanged(Editable editable) {
-               List<Contact> filterContact = mAdapter.getFilter(editable.toString());
+               //REVISIT : Use the presenter TempPresenterImpl to do any update to model.
+               //List<Contact> filterContact = mAdapter.getFilter(editable.toString());
            }
        });
-    }
-    public void onButtonPressed(Uri uri) {
-        if (mListener != null) {
-            mListener.onFragmentInteraction(uri);
-        }
     }
 
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
-//        if (context instanceof OnFragmentInteractionListener) {
-//            mListener = (OnFragmentInteractionListener) context;
-//        } else {
-//            throw new RuntimeException(context.toString()
-//                    + " must implement OnFragmentInteractionListener");
-//        }
     }
 
     @Override
@@ -127,16 +101,11 @@ public class TempFragment extends Fragment {
         mListener = null;
     }
 
-    /**
-     * This interface must be implemented by activities that contain this
-     * fragment to allow an interaction in this fragment to be communicated
-     * to the activity and potentially other fragments contained in that
-     * activity.
-     * <p>
-     * See the Android Training lesson <a href=
-     * "http://developer.android.com/training/basics/fragments/communicating.html"
-     * >Communicating with Other Fragments</a> for more information.
-     */
+    @Override
+    public void setPresenter(TempContract.Presenter presenter) {
+        this.presenter =presenter;
+    }
+
     public interface OnFragmentInteractionListener {
         // TODO: Update argument type and name
         void onFragmentInteraction(Uri uri);
