+++ app/src/main/res/layout/activity_incoming_call.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".IncomingCallActivity">
+    tools:context=".presenter.IncomingCallActivity">
 
     <TextView
         android:layout_width="wrap_content"
