+++ app/src/main/res/layout/fragment_temp.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".TempFragment">
+    tools:context=".view.TempFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <EditText
